requirements.txt
antlr4-python3-runtime==4.9.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
easyocr==1.7.1
filelock==3.16.0
filetype==1.2.0
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.9.0
gunicorn
idna==3.7
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
omegaconf==2.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.0.3
pillow==10.4.0
psutil==6.0.0
pytesseract
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pyparsing==3.1.4
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.44
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
sympy==1.13.2
tifffile==2023.7.10
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.92
ultralytics-thop==2.0.6
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
